+++ summary_result.h
@@ -12,7 +12,7 @@ typedef struct SUMMARY_STAT
   /* for calculating average */
   //  double lr_total;		/* sum of log(LR) */
   int scale;                    /* scale of the rescaled LR - to handle overflow */
-  int lr_count;			/* number of models for the (ld, theta) pair */
+  long long lr_count;			/* number of models for the (ld, theta) pair */
   double het_lr_avg;		/* average of HET log(LR) */
   double het_lr_total;		/* with heterogeneity - alpha, parameter */
   double het_lr_avg_orig;	/* average of HET log(LR) */
